Day 9/event_verifier/requirements.txt
langchain
google-generativeai
python-dotenv
PyMuPDF
pytesseract
Pillow
unstructured
